requirements.txt
# PNCT Container Query System - Python Dependencies

# Core Framework
fastapi>=0.109.0          # FastAPI web framework
uvicorn[standard]>=0.27.0  # ASGI server for FastAPI

# HTTP Client
httpx>=0.27.0             # Async HTTP client

# Data Validation
pydantic>=2.0.0           # Data validation using Python type annotations

# Environment Variables
python-dotenv>=1.0.0      # Load environment variables from .env file

# AI Agent (Gemini)
google-genai>=1.0.0       # Google Gemini AI SDK

# Temporal Workflows
temporalio>=1.8.0         # Temporal Python SDK for workflows

# Frontend
streamlit>=1.28.0         # Streamlit web framework for chat interface

# Testing (optional)
pytest>=7.4.0             # Testing framework
pytest-asyncio>=0.21.0    # Async testing support
